Implement cFS app generation and cFS handler generation. Refs #3.
Ogma must give users the ability to process FRET files and generate
monitors automatically.

This commit introduces commands and transformation functions to generate
NASA core Flight System (cFS) applications and message handlers from
Copilot monitors.

The implementation relies on a cFS application template, which could be
customized further.

The implementation is added as an extension to the ogma-core library.

The ogma-extra package is extended with functions needed to manipulate
directories and files.