Introduce library for C header parsing/AST representation. Refs #1.
Ogma needs to be able to read HEADER files and generate Copilot code
from them.

This commit implements a C parser and AST by means of the BNFC parser
generator. The grammar is missing one rule in type identifiers, left
commented out, due to a conflict, and extended with specific types for
integers and booleans.

The implementation is isolated in a separate library.