Introduce library for Copilot struct abstraction. Refs #1.
Ogma needs to be able to generate Copilot declarations for structs from
C struct definitions. Copilot's structs are defined similarly to C
structs, with a number of fields, each having a different type.

This commit introduces a type to represent the elements of the Copilot
language that must be first class and manipulated at meta-level.

The implementation is included as a separate library.