Introduce library to manipulate FRET Requirements DBs. Refs #2.
Ogma must give users the ability to process FRET files and generate
monitors automatically.

FRET files can be of two kinds: internal requirements DBs or Component
Specifications.

This commit introduces a frontend for FRET Requirement Databases.

The implementation is added as a separate library.

The ogma-extra package is extended with functions needed to manipulate
bytestrings.