Introduce library implementing conversion of C structs into Copilot structs. Refs #1.
The implementation of Ogma requires that users be able to generate the
necessary declarations in Copilot to work with C structs.

This commit introduces functions to convert from the abstract C AST to
Copilot structs.

The implementation is included as a core library that implements the
functions as IO but does not offer any user interface.