Introduce license and README at top level. Refs #4.
The license and the README should be accessible from the top level of
the repository so that they are easily accessible for those who access
the repo, and for code navigation tools that know how to treat these
files.

This commit introduces these two files as symbolic links to the
files with the same names in the ogma-cli directory.