Replace symbolic link for destination. Refs #5.
The LICENSE.pdf file at the top level uses a symbolic link to point to
the license in the ogma-cli directory.

Although github correctly identifies the file as the license, and also
that the file is a symbolic link, it does not give users a direct way of
visualizing the PDF from the link on the License option of the repo.

This commit replaces the symbolic link with the actual file pointed to
by it, which hopefully will help github locate the file and present it
to users for visualization.