Update installation instructions for cabal-install 3.6. Refs #6.
Installation with newer versions of cabal-install no longer uses
sandboxes.

This commit replaces the installation instructions for a set that is
working with cabal-install 3.6. The instructions should work with any
version of cabal after 2.4.

At present, there are two ways of installing Haskell packages with
cabal: using the v1-* and the v2-* installation commends.

The v2-* family is more modern, but it appears to make it hard for GHC
to find packages without having to write an additional cabal package
file for the Copilot examples that we expect our users to write.

This commit introduces instructions that use the v1-* family, but
adapted to the removal of sandboxes that occurred in Cabal 3.4.  The
packages are installed globally, which we expect will suit most users
who are not Haskell developers.

The installation requires installing alex and happy first, then BNFC and
copilot, and finally ogma. The order is deliberate, and due to the fact
that the v1-* commands do not seem to detect that BNFC uses alex and
happy and they should be installed first.