Hide (==) from Prelude in generated Copilot code. Refs #7.
The code currently being generated by Ogma uses the stream equality
operator (==) without hiding the one from the Prelude. Because this is
not being overloaded, the compiler cannot tell which one is being
referred to.

This commit hides (==) in the explicit Prelude import in the code
generated by Ogma from FRET Component Specifications.