Add extra parenthesis to output. Refs #8.
The code being generated by Ogma for FRET Component Specifications
assumes that the operator precedence in FRET and Copilot is the same,
which is not necessarily true for all operators. This can lead to type
errors or difference in meaning in the generated code and the intended
code.

This commit adds parentheses in the output around applications of binary
operators to boolean expressions.