Handle int type from FRET Component specifications. Refs #9.
Some of the FRET Component Specifications being generated by FRET use
the type int to refer to integers. Currently, Ogma is unable to handle
this type, which is printed verbatim to the output Copilot file and
interpreted by the Haskell compiler as a type variable instead of a
concrete type.

This commit modifies the converter to assume int to be another name for
Int64.