Constrain version of aeson in dependencies2. Refs #10.
Currently, Ogma's cabal files do not specify a version constraint on
Aeson, but the interface has changed in recent versions and it fails to
compile.

This commit modifies the cabal files of packages that use Aeson to
require a version prior to 2, which introduced the breaking change to
the interface.