Swap order of time bounds in past-time MTL operators. Refs #11.
The time bounds passed in Copilot to past-time MTL operators
(alwaysBeen, eventuallyPrev) are swapped. They should contain the lower
number first, but they contain the higher first. This is a direct
translation of the FRET file, which uses that notation. They should be
swapped.

This commit swaps the order of the time bounds passed to alwaysBeen
and eventuallyPrev in the CoCoSpec generator.