Simplify unnecessarily structured productions. Refs #21.
The SMV grammar gives names to numeric expressions between bold/italics
HTML tags, even though they have no significance and act more as
parenthesis.

The BNFC format that the grammar is written in provides a mechanism to
parse but discard these extra nesting without affecting the AST, by
using an underscore (_) as the constructor name for those productions.

This commit replaces the constructors in those productions for
underscores, which removes the old construtors from the AST. The Copilot
backend is therefore adapted so it does not pattern match on those
removed data constructors.