Introduce flags to control numeric types in Copilot code. Refs #12.
Ogma currently defaults all ints from FRET components to Int64, and all
reals to Float. This is not always suitable, since the code needs to be
integrated with systems that may use a different representation for
integers and reals.

This commit introduces two new flags to control the representation used
for FRET's real and int types.