Wrap command options in additional type that is easier to extend. Refs #12.
As we continue to introduce options in commands, the number of arguments
that must be passed around from the top-level interface (CLI) to the
actual transformers begins to grow and becomes harder to manage.

This commit introduces a type to enclose all arguments for the FRET
Requirements Database to Copilot transformation command, and the
underlying transformation functions. Its use will facilitate introduce
new flags and options without affecting the function's interface.