Include missing imports from ICAROUS in generated cFS app. Refs #26.
The generated cFS app by default includes a monitor from the ICAROUS
system (https://nasa.github.io/icarous/). However, the C files does not
include the ICAROUS headers, and the CMakeLists.txt does not refer to
the ICAROUS app directories.

This commit adds the missing includes to the C code and adjusts the
CMakeLists.txt to find the necessary files in a standard ICAROUS
installation.