Fix error in installation note. Refs #27.
The README claims:

"these instructions have been tested with GHC 8.10 and Cabal versions up to 3.6"

However, that is not true. The instructions do not work for cabal 3.6,
but ogma does.

This commit modifies that claim to state that Ogma is tested with more
modern versions of GHC/cabal, but the instructions may differ.