Fix MTL operator names. Refs #30.
The Copilot code generated by Ogma uses the same temporal operators for
PTLTL and MTL. However, Copilot's interface for those operators is not
the same: although the names coincide, the number (and type) of
arguments do not. This results in (sometimes) invalid code.

This commit introduces an auxiliary function to determine the temporal
operator names specifically for MTL properties.