Briefly introduce FRET, add image of FRET and X-Plane. Refs #16.
To help users unfamiliar with FRET, this commit briefly introduces the
project, and includes two images: one of FRET, showing a requirement in
the editor, and one of X-Plane, showing the same requirement being
monitored in the simulator.

This commit also includes minor fixes to the FRET explanation suggested
by Anastasia Mavridou from the FRET team.