ogma-extra:Data.ByteString.Extra: reduce indent. Refs #38.
The implementation of ogma-extra:Data.ByteString.Extra.safeReadFile is
aligned inconsistently with similar functions like
ogma-extra:Data.String.Extra.safeReadFile. The indentation level of a
sub-expression is not the same in both, even though they are built the
same way.

This commit reduces the indentation in ByteString to match String. The
decision is due to the latter using two spaces only, which is also
consistent with other indentation decisions.