ogma-cli: Make Hlint ignore suggestion. Refs #51.
We purposefully use (id <$>) because conceptually there is a
transformation taking place, but in this case it would be the identity.
This is (almost) a coincidence, and may not be like that in the future.

We indicate this fact with the use of a transformation function in the
code, although such function in this case does nothing.

This commit adds an annotation so that hlint ignores that specific
suggestion, and a comment to explain the decision.