ogma-language-lustre: Introduce package to process Lustre expressions. Refs #267.
There is a need for Ogma to be able to parse Lustre expressions.
Allowing for Lustre would increase the versatility of Ogma, and would
allow engineers to use languages they may be familiar with and leverage
Ogma's capabilities.

This commit renames the ogma-language-cocospec, which uses the same
syntax (with specific temporal logic operators), to
ogma-language-lustre, replacing all references and filenames
accordingly.