ogma-cli: Update to support boolean expressions in the Lustre language. Refs #267.
There is a need for Ogma to be able to parse Lustre expressions.
Allowing for Lustre would increase the versatility of Ogma, and would
allow engineers to use languages they may be familiar with and leverage
Ogma's capabilities.

A prior commit has introduced a package ogma-language-lustre that
substitutes an existing package in the codebase.

This commit replaces all occurrences of the package by the old name, as
well as all pertinent filenames and expression formats, to use the name
Lustre or lustre instead, as applicable.