ogma-language-fret-cs: Update to work with Aeson 2.0. Refs #55.
GHC 9.2 is the current default using macports. Ogma does not currently
support it because one of the dependencies (aeson) must be upgraded.

This commit adjusts the code of ogma-language-fret-cs to Aeson 2's new
API. Two imports are replaced by Aeson modules, so two packages are no
longer required by this library and are removed from the cabal file.