ogma-core: Update license in cabal file to OtherLicense. Refs #62.
The license indicated in the license field of the cabal files is
AllRightsReserved. Using such a license prevents packages from being
uploaded to hackage.

This commit updates the license field to OtherLicense, which is
compatible with our actual license, and allows the package to be
uploaded to hackage, facilitating installation for our users.