ogma-language-c: Add src/.keep file to distributable package. Refs #65.
Packages that only consist of automatically generated files cannot be
uploaded to hackage because the source directory where the files are
being located is not included by cabal in the tar.gz during the dist
command.

This commit adds the empty src/.keep file to the extra-source-files
field so that the directory structure is included in the distributable
package.