ogma-core: Add fuzz test generation to ROS 2 backend. Refs #280.
It would be useful to be able to leverage monitors to perform fuzz
testing on ROS 2 systems, where the monitors could be used as oracles.

This commit extends the ROS 2 backend to populate the application data
with information that is leveraged by a new test_requirements ROS 2
package in order to feed random data to ROS 2 topics and use the
monitors as oracles.