ogma-cli: Add flags to control ROS 2 fuzz test generation. Refs #280.
A prior commit has extended the ROS 2 backend to generate code to
perform fuzz testing on ROS 2 systems, where the monitors being
generated are used as test oracles.

This commit adds two flags to the ROS 2 backend to control the
generation of tests for ROS 2. Specifically, the flags allow controlling
which variables (or topics) the random data is generated for, and what
ROS 2 nodes are turned on during testing (apart from the Copilot
monitors and the tests).