ogma-extra: Mark package as uncurated. Refs #74.
Cabal packages in Ogma are not using the flag x-curation to indicate to package
maintainers that our packages should not be modified by them.

Because this is a NASA project, we want to make sure that users obtain exactly
what we publish, unmodified by anyone external to our project.

This commit sets modifies the cabal package to include x-curation: uncurated,
signaling our preferences to the Hackage maintainers.