ogma-core: Version bump (1.0.8). Refs #81.
We choose to bump the version constraints for all dependencies at once to keep
them in sync, since it will be much easier than trying to track the version
number of each dependency inside the Ogma project.

In this case, the change is needed also because this version of ogma-core needs
definitions that will be published in ogma-language-smv and
ogma-language-cocospec.