ogma-language-fret-cs: Improve parsing error messages. Refs #96.
Ogma relies on an existing library for error message reporting of JSON
files. That library, by default, provides very plain error messages and
the line and column number of the error cannot be customized or
contextualized. This causes error messages to seem to apply to a line
and column in the file, when they really apply to the line and column of
a specific requirement (which may be hard to indentify).

This commit modifies the JSON parser for component specification files
so that, if a parsing error in a temporal logic formula is found, the
parser reports the name of the requirement that failed to parse.