ogma-core: Translate ZtoPre and YtoPre to Copilot. Refs #86.
Ogma cannot handle CoCoSpec boolean specifications containing the unary
operators ZtoPre or YtoPre. This makes some of the specifications produced by
FRET not parseable.

The CoCoSpec grammar has been extended with two new unary operators ZtoPre and
YtoPre.

This commit updates the translator from CoCoSpec to Copilot so that it assigns
a Copilot construct to those new unary operators. The component spec and
requirements db translators are updated to include the necessary Copilot
definitions so that the code produced compiles without errors.