ogma-language-smv: Support MTL operators with number ranges. Refs #101.
The current parser for the SMV language does not support MTL operators with
number ranges as arguments to MTL operators. The only syntax currently
supported is using <MTLOperator>[<ComparisonOperator><Number>], where is a
numeric comparison operator like <=, <, and so on.

This limitation makes some of the component specifications produced from FRET
not parseable with Ogma.

This commit extends the grammar to support MTL operators with number ranges.
The prior syntax is also supported and kept as is.