ogma-core: Modify reqs DB backend to allow customizing target filename. Refs #80.
Ogma does not allow users to specify the name of the target file that should be
used in applications produced from FRET files. Without that extra facility, it
becomes necessary to rename and modify the Copilot/Haskell (or C) files
produced, which detracts from a smooth UX process.

This commit modifies the FRET requirements DB backend to customize the name of
the target C filenames based on an argument to the backend.