ogma-cli: Fix rendering of quotes in URLs in package description. Refs #72.
The description in the cabal file contains two entries with double quotes in
the tiles of links. Those double quotes are not displayed correctly on hackage:
instead of appearing as ", they are shown as &quot;.

This commit moves the quotes outside the link text so that hackage will render
them without escaping them.