ogma-core: Replace all mentions of the Sample App. Refs #105.
The cFS application produced by Ogma includes multiple mentions of the Sample
App. This is because template used is based on cFS' Sample Application. Keeping
those references there is misleading, since this is not the sample app, but a
monitoring app.

This commit replaces all mentions of the sample app to the Copilot app.