ogma-cli: Document cabal update step in README. Refs #100.
The installation instructions in the README do not mention the cabal update
step. This is confusing users unfamiliar with Haskell, since it's not obvious
that cabal should have to be updated before installing Ogma.

This commit adds an update step prior to installing any tools in the
installation instructions in the README. We use the v1 variant of the command,
for consistency with the rest of the instructions.