ogma-spec: Introduce library to represent ogma specifications. Refs #115.
Ogma currently has no standard representation for specifications, and
instead relies on a representation that is tied to the input format from
which the spec was read.

This commit introduces a new library to represent generic Ogma
specifications. The specifications are parameterized by the type of the
requirement expressions.

This will allow us to use the same internal representation to hold data
obtained from different kinds of input files (e.g., JSON, XML).