ogma-language-jsonspec: Introduce library to parse JSON specs. Refs #115.
The JSON parsing functions currently included with Ogma are very
specific to two formats supported, and cannot be easily generalized.
This limits our ability to adapt to changes and to different formats.

This commit introduces a new library to parse Ogma specifications from
JSON files. The parsing functions are parameterized by the functions
that parse requirement sub-expressions, allowing for further
customization.