ogma-language-jsonspec: Make some fields in JSON format spec optional. Refs #122.
Not all kinds of input files that use the JSON parsing library will
provide all information. In some, information is implicit or simply not
available, yet we should be able to deal with such files.

This commit modifies the JSON spec parsing library so that some fields
can be optional. In general, we assume we'll still have the essential
fields (requirements, IDs, specs), but not all auxiliary information.