ogma-core: Fix missing stream name substitution. Refs #120.
When using property names as arguments to trigger, the Spec2Copilot
backend does not apply the same variable/name substitution that was
applied in their definition, thus referring in the call to trigger to an
undefined argument.

This commit modifies the generation of trigger names in Copilot to be
able to apply the same name subsitution map, so that the stream name is
resolved.