ogma-core: Use generalized JSON parser for DB Spec. Refs #122.
The JSON parsing functions currently used by ogma-core to parse
requirement DBs are very specific to the formats supported.

This commit replaces the existing FRET DB parser and converter by one
that is based on the new generic JSON-based parsing library. The old
module implementing the ad hoc converter, as well as the dependency on
the ad hoc parser, are removed.