ogma-language-xmlspec: Introduce library to parse XML specs. Refs #202.
There is a need in Ogma to support XML-based input files, since this is
produced by many standard tools in the industry. Like with JSON, we want
users to be able to customize the input format via a command-line flag,
so that users can work with XML-based files whose format is previously
unknown to Ogma without having to modify the tool.

This commit introduces a new library to parse Ogma specifications from
XML files. The parsing functions are parameterized by the functions that
parse requirement sub-expressions, allowing for further customization,
The sub-expression parser may execute an IO-action, enabling the use of
external commands and/or LLMs to pre-process expressions. The
specification of the XML format is readable, which will also enable
parsing a format specification from an input file (either a known file
distributed with Ogma or a user-provided input file).