ogma-core: Add support for XML files to standalone backend. Refs #202.
There is a need in Ogma to support XML-based input files, since this is
produced by many standard tools in the industry. Like with JSON, we want
users to be able to customize the input format via a command-line flag,
so that users can work with XML-based files whose format is previously
unknown to Ogma without having to modify the tool.

A prior commit has introduced a library to parse Ogma specifications
from XML files.

This commit extends ogma-core to accept input files in XML, and to treat
any configuration file containing an XMLFormat specification as format
specification for an XML-based input file.