ogma-core: Use same handler name in FPrime/ROS and Copilot. Refs #130.
The Copilot, FPrime and ROS backends do not use the same name for the
handler names. In one case, an extra "prop" is added to the name, which
makes the resulting Copilot code not directly usable from FPrime / ROS
without edits.

This commit modifies the FPrime and ROS backends so that the handler
names do not contain the additional "prop", making them match the
Copilot backend.