ogma-cli: Introduce CI test job. Refs #139.
Ogma doesn't currently rely on Travis CI for testing, because Travis is
setup to run on all NASA repositories and we run out of cycles very
quickly. We need to automate the testing process as much as possible
regardless, and follow our own guide for compliance with NPR7150.2D but
using Github instead of Travis CI.

This commit introduces a Github Actions file that builds the source code
in every push action and with every pull requests.