ogma-core: Add dockerfile. Refs #136.
Ogma doesn't currently generate a Dockerfile with the ROS packages. That
makes it harder for users to know how to use the produce of Ogma.

This commit adds a Dockerfile to the ROS template used by Ogma, and to
the list of files installed with Ogma. The dockerfile simply compiles
the monitoring application inside Space ROS.