ogma-core: Map float and double to the same types in C++. Refs #138.
The ROS backend is, by default, assuming that the data carried by
std_msgs::msg::Float32 is of type float32, when it appears to be just a
float. The same is happening for Float64, float64 and double. The code
currently produced for those cases does not compile.

The type mappings are described in
http://docs.ros.org/en/humble/Concepts/Basic/About-Interfaces.html,
which states that, in C++, float and double should be used.

This commit modifies the ROS backend so that it expects global variables
of input types `float` and `double` in the variable DB to have the types
`float` and `double`, respectively, in the generated C++ code.