ogma-cli: Add a float and a double input variable to ROS example. Refs #138.
This commit modifies an example to use also an input variable of type
float and one of type double, both of which were failing in the ROS
backend.

The example in question is currently being used by a CI script to test
the ROS backend, making the new example serve as a regression test.