ogma-cli: Document format of variable DB. Refs #143.
The README does not explain the structure that the variable DB files for
ROS should have. Because ogma doesn't give much information to users
when a variable cannot be found in the var DB, users do not know what to
put in each field in the variable DB, and what effect it will have in
the program.

This commit adds a subsection to the ROS section of the README
explaining the structure and format of the variable DB.