ogma-core: Use template expansion to generate cFS monitoring application. Refs #157.
The cFS application generation backend has a fixed template that it uses
to generate the cFS application. That template does not fit all use
cases, so we are finding users heavily modifying the output (which is
hard to keep up with when there are changes), and or not using ogma
altogether for that reason.

This commit modifies the ogma-core cfs command to use mustache to
generate the cFS monitoring application template using variable
expansion.