ogma-core: Enable customizing template directory in cFS app generation. Refs #157.
The cFS application generation backend has a fixed template that it uses
to generate the cFS application. That template does not fit all use
cases, so we are finding users heavily modifying the output (which is
hard to keep up with when there are changes), and or not using ogma
altogether for that reason.

A recent commit introduced the ability to use mustache to expand
variables in a template. This commit modifies the cfs command to accept
an additional argument that points to a user-provided directory with a
custom template.