ogma-core: Fix incorrect path when using Space ROS humble-2024.10.0. Refs #158.
The new Space ROS no longer has packages in the user's home. Instead,
the Space ROS installation has been moved to /opt. This makes the
Dockerfile produced by the ROS backend fail, since it sources a
setup.bash file that no longer exists.

This commit fixes the version of Space ROS to be humble-2024.10.0, and
adjusts a source command to the correct location under that version.