ogma-core: Use template expansion to generate ROS monitoring package. Refs #162.
The ROS package generation backend has a fixed template that it uses to
generate the ROS nodes. That template does not fit all use cases, so we
are finding users heavily modifying the output (which is hard to keep up
with when there are changes), and or not using ogma altogether for that
reason.

This commit modifies the ogma-core ros command to use mustache to
generate the ROS monitoring package via a template and variable
expansion. We adapt the template to also use those variables, and we
modify the cabal file to include the new files as data files that are
copied over during installation.