ogma-core: Enable customizing template directory in ROS package generation. Refs #162.
The ROS package generation backend has a fixed template that it uses to
generate the ROS package. That template does not fit all use cases, so
we are finding users heavily modifying the output (which is hard to keep
up with when there are changes), and or not using ogma altogether for
that reason.

A recent commit introduced the ability to use mustache to expand
variables in a template. This commit modifies the ros command to accept
an additional argument that points to a user-provided directory with a
custom template.