ogma-cli: Provide CLI argument to pass an auxiliary template source directory to ROS app command. Refs #162.
The ROS package generation backend has a fixed template that it uses to
generate the ROS package. That template does not fit all use cases, so
we are finding users heavily modifying the output (which is hard to keep
up with when there are changes), and or not using ogma altogether for
that reason.

A recent commit introduced into ogma-core the ability to use a custom
provided template and expand variables using mustache.

This commit exposes that new parameter to the user in the CLI.