ogma-cli: Update README to document ROS package template customization. Refs #162.
The ROS package generation backend used a fixed template to generate the
ROS package. That template did not fit all use cases, and users are
modifying the output (which is hard to keep up with when there are
changes), and or not using ogma altogether for that reason.

Prior commits have expanded the command to allow for customization of
the template using a user-provided directory and expanding variables in
the template using mustache.

This commit documents the new feature in the README.