ogma-core: Use template expansion feature from ogma-extra in cFS backend. Refs #162.
The ROS backend used a fixed template to generate the ROS package. That
template does not fit all use cases, and users are heavily modifying the
output (which is hard to keep up with when there are changes), and or
not using ogma altogether for that reason. To make it possible for users
to customize the ROS output, we want to support using mustache templates
for ROS packages, like we did with the cFS backend.

A prior commit has copied the functionality of template generation and
expansion using mustache from ogma-core's cFS backend to ogma-extra.
This commit makes the cFS backend depend on that feature on ogma-extra,
to avoid duplication.