ogma-cli: Introduce standalone command. Refs #170.
The CLI of Ogma lacks uniformity: some commands are driven by the input
format, others by the output format (e.g., ROS, cFS). To make Ogma
easier to use, we should provide an interface that is driven by one or
the other, but not both.

A prior commit has introduced a new standalone command, at core level,
that produces an individual Copilot specification from an input file.

This commit exposes that command to the user via the CLI.