ogma-language-jsonspec: Make JSONFormat readable. Refs #170.
To make Ogma more easily extensible, we want to make it possible to
store the format specifications in individual files, so that they do not
have to be included with the code and they can be provided by users.

This commit makes a JSONFormat readable, so that it can be easily read
from a file.