ogma-core: Remove redundant commands. Refs #170.
The CLI of Ogma lacks uniformity: some commands are driven by the input
format, others by the output format (e.g., ROS, cFS). To make Ogma
easier to use, we should provide an interface that is driven by one or
the other, but not both.

A prior commit has introduced a new standalone command that can be used
to produce individual Copilot specs, and redundant commands have been
removed from the CLI. This commit removes the corresponding redundant
core commands.