ogma-cli: Adjust tests to use new standalone command. Refs #170.
The CLI of Ogma lacks uniformity: some commands are driven by the input
format, others by the output format (e.g., ROS, cFS). To make Ogma
easier to use, we should provide an interface that is driven by one or
the other, but not both.

A prior commit has introduced a new standalone command that replaces two
prior existing commands.

This commit adjusts the CLI package tests to use the new standalone
command.