ogma-cli: Update README to demonstrate robotics. Refs #172.
The README shows an image of a monitor in X-Plane, but nothing related
to robotics, even though it is one of the main use cases for Copilot.
This should be demonstrated more clearly.

This commit updates the README to show a video of Ogma-generated
monitors running on a ROS system, integrated into Gazebo.