ogma-core: Correct name in documentation. Refs #176.
The documentation of ogma-core/src/Language/Trans/SMV2Copilot.hs does
not refer to the boolean expressions as those in the SMV language, but
incorrectly refers to them as FRET expressions, which was old
documentation left over after renaming the module.

This commit renames all occurrences to SMV in the module documentation.