ogma-cli: Update README to document F' component template customization. Refs #185.
The F' backend uses a fixed template to generate the F' component. That
template does not fit all use cases, so we are finding users heavily
modifying the output (which is hard to keep up with when there are
changes), and or not using ogma altogether for that reason.

Prior commits have expanded the command to allow for customization of
the template using a user-provided directory and expanding variables in
the template using mustache.

This commit documents the new feature in the README.