ogma-core: Use template expansion to generate standalone monitoring application. Refs #189.
The standalone backend uses a fixed template to generate the Copilot
monitor. That template does not fit all use cases, so we are finding
users heavily modifying the output (which is hard to keep up with when
there are changes), or not using ogma altogether for that reason.

This commit modifies the ogma-core standalone command to use mustache to
generate the Copilot monitor via a template and variable expansion. We
introduce a new template that uses variables, and we modify the cabal
file to include the files that make up the default template as data
files that are copied over during installation.

To be able to generate the files, we need to introduce a target
directory option for the standalone command.