ogma-core: Adjust tests to provide target directory to standalone command. Refs #189.
The standalone backend uses a fixed template to generate the Copilot
monitor. That template does not fit all use cases, so we are finding
users heavily modifying the output (which is hard to keep up with when
there are changes), or not using ogma altogether for that reason.

A prior commit introduced, in the standalone command, support for custom
templates. Unlike the prior implementation of the standalone command,
which printed the output to standard output, the new interface puts the
result in a file (or several files) included with the template.  To
implement this new functionality, we have added an argument to specify a
target directory for the generated standalone application,
where the files will be copied to.

This commit modifies the tests for the standalone command to provide a
temporary directory as destination. We default to the temporary
directory of the host system. Since this is only used for testing and
not during normal operation, and the ogma implementation is open source,
we consider this safe from a security standpoint.