ogma-core: Enable customizing template directory in standalone command. Refs #189.
The standalone backend uses a fixed template to generate the Copilot
monitor. That template does not fit all use cases, so we are finding
users heavily modifying the output (which is hard to keep up with when
there are changes), or not using ogma altogether for that reason.

A prior commit introduced, in the standalone command, the ability to use
mustache to expand variables in a template.

This commit modifies the standalone command to accept an additional
argument that points to a user-provided directory with a custom
template.