ogma-cli: Add CLI argument to standalone command to set auxiliary template source directory. Refs #189.
The standalone backend uses a fixed template to generate the Copilot
monitor. That template does not fit all use cases, so we are finding
users heavily modifying the output (which is hard to keep up with when
there are changes), and or not using ogma altogether for that reason.

A recent commit introduced into ogma-core the ability to use a custom
template and expand variables using mustache.

This commit exposes that new parameter to the user in the CLI.