ogma-core: Adjust tests to provide template directory to standalone command. Refs #189.
The standalone backend uses a fixed template to generate the Copilot
monitor. That template does not fit all use cases, so we are finding
users heavily modifying the output (which is hard to keep up with when
there are changes), or not using ogma altogether for that reason.

A prior commit introduced, in the standalone command, support for custom
templates and, with it, the ability to pass a custom directory
containing a standalone application template.

This commit modifies the tests for the standalone command to provide no
custom template directory, to use the default included with Ogma.