ogma-cli: Add repository information to cabal package. Refs #148.
The cabal packages do not contain pointers to the repo using the
specific entry for this in Cabal, neither for the source, nor to file
bug reports. Although such information is not, strictly speaking,
necessary, it helps direct users without having to "jump around" to find
out where they need to go.

This commit adds both the bug-reports link and the source repository
information to the cabal package.