ogma-language-jsonspec: Extend JSON spec parser to use an IO action to parse properties. Refs #197.
There is a need in Ogma to make the frontend more versatile, and to
extend it to support languages unknown to Ogma without users having to
modify the code of Ogma itself. To that end, we want to empower users
with the ability to use a custom command to transform individual
properties into a format that Ogma understands, and have Ogma call that
external command on demand.

Because calling an external command will be an IO action, any auxiliary
sub-parsers used by Ogma should be IO actions, rather than pure
functions.

This commit modifies the parser for JSON specifications so that it takes
as argument an IO action to parse individual properties.