ogma-core: Adapt ROS, F' backend to use IO action to parse individual properties. Refs #197.
There is a need in Ogma to make the frontend more versatile, and to
extend it to support languages unknown to Ogma without users having to
modify the code of Ogma itself. To that end, we want to empower users
with the ability to use a custom command to transform individual
properties into a format that Ogma understands, and have Ogma call that
external command on demand.

Prior commits have extended the JSON parser to use an external IO action
to parse individual properties, rather than pure functions, which is
suitable when those actions represent calls to external processes.

This commit modifies the ROS and F' backends to wrap parser functions
into IO actions, as required by the modified JSON specification parser.