ogma-cli: Provide ability to preprocess properties via external command. Refs #197.
There is a need in Ogma to make the frontend more versatile, and to
extend it to support languages unknown to Ogma without users having to
modify the code of Ogma itself. To that end, we want to empower users
with the ability to use a custom command to transform individual
properties into a format that Ogma understands, and have Ogma call that
external command on demand.

Prior commits have extended the standalone backend with the capability
of taking an auxiliary command as argument and running it to transform
individual properties.

This commit exposes that new argument to the user in the CLI.