ogma-core: Introduce XML format specifications. Refs #202.
There is a need in Ogma to support XML-based input files, since this is
produced by many standard tools in the industry. Like with JSON, we want
users to be able to customize the input format via a command-line flag,
so that users can work with XML-based files whose format is previously
unknown to Ogma without having to modify the tool.

A prior commit has introduced the capability to use XML files with
ogma-core.

This commit adds four XML format specifications based on two standard
formats used in known MBSE tools.