ogma-core: Enable using user-provided file as format definition spec. Refs #200.
Ogma allows users to select the format used for the JSON file, but only
known formats can be selected. Users should be able to pass a local file
containing a JSON format specification as an argument.

This commit modifies the standalone backend so that, if a file exists
with the name of the format selected, or if the format name contains a
path separator, then it is treated as a local file. Otherwise, Ogma
finds the file in the package's data directory.