ogma-core: Rename options in FPrime backend for versatility. Refs #204.
The FPrime backend used to be specialized for one file format as input,
which was reflected in the name of the options expected by the backend.

For consistency across backend, and to facilitate the introduction of
support for different kinds of file formats in the future, we rename the
backend's options to be more format-agnostic.