ogma-cli: Rename options in FPrime backend for versatility. Refs #204.
The FPrime backend used to be specialized for one file format as input,
which was reflected in the name of the options expected by the backend.

A prior commit has renamed the options expected by Ogma's backend in
ogma-core to make them more format-agnostic, in alignment with the plan
to support different kinds of file formats.

This commit performs the same change for users in the CLI.,