ogma-core: Add support to ROS backend for different property formats. Refs #204.
The standalone backend supports cocospec, smv and literal Copilot
expressions, but that capability is not yet available in the ROS
backend.

This commit adds that capability by introducing a notion of expression
pairs or expression input file handlers parser can use to parse
expressions or properties contained in input files.