ogma-cli: Make ROS backend able to handle different formats. Refs #204.
The standalone backend supports JSON and XML, which can be further
confirured in the command line, but that capability is not yet available
in the ROS backend in the command line interface.

A prior commit has expanded the ROS backend in ogma-core with support
for the aforementioned capability.

This commit exposes the new capability to the user in the CLI.